requirements.txt
pywin32
pycryptodome
pandas
openpyxl